requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
langchain-openai
chromadb
tiktoken
python-multipart
streamlit
pydantic-settings
python-dotenv
pypdf
sentence-transformers
openai
requests
rank-bm25
aiofiles
jinja2
